Upd. Auth. Refactoring auth handlers.
