+++ inc/spbc-admin.php
@@ -672,147 +672,6 @@ function spbc_validate_secure_cookies($cookies_arr)
     }
 }
 
-/**
- * Add option to the profile page.
- *
- * @param WP_User $wp_user
- */
-function spbc_2fa__SelfUserProfileEdit($wp_user)
-{
-    global $spbc;
-
-    if (spbc_is_user_role_in($spbc->settings['2fa__roles'], $wp_user)) {
-        $type2fa = get_user_meta($wp_user->ID, 'spbc_2fa_type', true);
-        $button  = '';
-
-        if ($type2fa === 'google_authenticator') {
-            $type2fa_label = esc_html__('Google authenticator', 'security-malware-firewall');
-            // Disable G2FA can only admin and self-edited user.
-            if (current_filter() === 'show_user_profile' || current_user_can('manage_options')) {
-                $button = '<button class="button" id="spbc-g2fa-disable" data-user-id="' . sanitize_key((string) $wp_user->ID) . '">' . esc_html__('Disable Google Authenticator', 'security-malware-firewall') . '</button>';
-                $button .= '<br><em>' . esc_html__('To disable the Google authentication click the button above or reset the password to the account. The two-factor authentication will be switched to Email. Or you can disable it directly on the page of the WordPress site profile.', 'security-malware-firewall') . '</em>';
-            }
-        } else {
-            $type2fa_label = esc_html__('Email', 'security-malware-firewall');
-            // Enable G2FA can only self-edited user.
-            if (current_filter() === 'show_user_profile') {
-                $button = '<button class="button" id="spbc-g2fa-enable">' . esc_html__('Enable Google Authenticator', 'security-malware-firewall') . '</button>';
-            }
-        }
-
-        echo '<h2>' . $spbc->data["wl_brandname"] . esc_html__(' 2FA', 'security-malware-firewall') . '</h2>';
-        ?>
-        <table class="form-table">
-            <tr id="spbc-2fa-type">
-                <th>
-                    <label for="spbc-2fa-type"><?php esc_html_e('2FA type', 'security-malware-firewall'); ?></label>
-                </th>
-                <td>
-                    <p><?php echo esc_html__('Two-factor authentication (2FA) type:', 'security-malware-firewall') . ' '; ?>
-                        <strong><?php echo $type2fa_label; ?></strong></p>
-                    <p><?php echo $button; ?></p>
-                </td>
-            </tr>
-        </table>
-        <div id="spbct-google-qr-code" class="spbc_hide" style="padding: 0 15px;text-align:center;">
-            <p><?php esc_html_e('Please scan this with the Google Authenticator App.', 'security-malware-firewall'); ?></p>
-            <div id="spbct-google-qr-code-img"></div>
-            <p><?php esc_html_e('Enter Google Authenticator code.', 'security-malware-firewall'); ?></p>
-            <input name="spbct-google-qr-code" type="text"/>
-        </div>
-        <?php
-    }
-}
-
-/**
- * Getting Google QR
- * AJAX action.
- */
-function spbc_2fa__GetGoogleQrCode()
-{
-    spbc_check_ajax_referer('spbc_secret_nonce', 'security');
-    $user_obj = wp_get_current_user();
-
-    if (property_exists($user_obj, 'ID')) {
-        $user_token = get_user_meta($user_obj->ID, 'spbc_g2fa_token', true);
-        $ga = new GoogleAuthenticator();
-
-        if (!$user_token) {
-            $user_token = $ga->generateSecret();
-            update_user_meta($user_obj->ID, 'spbc_g2fa_token', $user_token);
-        }
-
-        $url        = sprintf("otpauth://totp/%s?secret=%s", $user_obj->get('user_login') . '(' . get_site_url() . ')', $user_token);
-        $encoder    = 'https://chart.googleapis.com/chart?chs=200x200&chld=M|0&cht=qr&chl=';
-        $qrImageURL = sprintf("%s%s", $encoder, urlencode($url));
-        $qr         = '<img style="border:0;padding:10px;width:200px;height:auto;" src="' . $qrImageURL . '" alt="Google authenticator QR code"/>';
-
-        wp_send_json_success(array('img' => $qr, 'code' => $user_token));
-    }
-
-    wp_send_json_error(esc_html__('Error: Current user is undefined.', 'security-malware-firewall'));
-}
-
-/**
- * Checking 6 digit Google 2fa code.
- * AJAX action.
- */
-function spbc_2fa__CheckGoogleCode()
-{
-    spbc_check_ajax_referer('spbc_secret_nonce', 'security');
-
-    if (isset($_POST['code']) && preg_match('/^\d{6}$/', trim($_POST['code']))) {
-        $user_obj = wp_get_current_user();
-
-        if (property_exists($user_obj, 'ID')) {
-            $user_token = get_user_meta($user_obj->ID, 'spbc_g2fa_token', true);
-
-            if ($user_token) {
-                $ga   = new GoogleAuthenticator();
-                $code = $ga->getCode($user_token);
-
-                if (trim($_POST['code']) === $code) {
-                    if (update_user_meta($user_obj->ID, 'spbc_2fa_type', 'google_authenticator')) {
-                        wp_send_json_success();
-                    }
-                    wp_send_json_error(esc_html__('Error: 2FA type not updated.', 'security-malware-firewall'));
-                }
-
-                wp_send_json_error(esc_html__('Error: The code not match.', 'security-malware-firewall'));
-            }
-
-            wp_send_json_error(esc_html__('Error: User token error.', 'security-malware-firewall'));
-        }
-
-        wp_send_json_error(esc_html__('Error: Current user is undefined.', 'security-malware-firewall'));
-    }
-
-    wp_send_json_error(esc_html__('Error: The code is incorrect.', 'security-malware-firewall'));
-}
-
-/**
- * Disable Google 2fa.
- * AJAX action.
- */
-function spbc_2fa__DisableGoogle2fa()
-{
-    spbc_check_ajax_referer('spbc_secret_nonce', 'security');
-
-    if (isset($_POST['user_id'])) {
-        $user_obj = spbc_get_user_by('id', sanitize_key($_POST['user_id']));
-
-        if (is_object($user_obj) && property_exists($user_obj, 'ID')) {
-            if (spbc_2fa__Google2fa_replace_meta($user_obj)) {
-                wp_send_json_success();
-            }
-
-            wp_send_json_error(esc_html__('Error: 2FA type not updated.', 'security-malware-firewall'));
-        }
-    }
-
-    wp_send_json_error(esc_html__('Error: Current user is undefined.', 'security-malware-firewall'));
-}
-
 function spbc_get_authorized_admins($direct_call = false)
 {
     if ( ! $direct_call) {
