+++ lib/CleantalkSP/Updater/UpdaterScripts.php
@@ -1142,4 +1142,9 @@ class UpdaterScripts
         $spbc->settings['wp__dashboard_widget__show'] = 1;
         $spbc->save('settings');
     }
+        
+    public static function updateTo_2_120_0() //phpcs:ignore PSR1.Methods.CamelCapsMethodName.NotCamelCaps
+    {
+        delete_option('spbc_login_attempts');
+    }
 }
